InvoiceMaturity data type

+++ IdokladSdk/Models/IssuedDocumentTemplate/Patch/IssuedDocumentTemplatePatchModel.cs
@@ -66,7 +66,7 @@ namespace IdokladSdk.Models.IssuedDocumentTemplate.Patch
         /// <summary>
         /// Gets or sets invoice maturity (in days).
         /// </summary>
-        public short? InvoiceMaturity { get; set; }
+        public int? InvoiceMaturity { get; set; }
 
         /// <summary>
         /// Gets or sets fixed exchange rate.
